// infrastructure/microservices/task-microservice/src/Domain/enums/ErrorCode.ts
export enum ErrorCode {
  NONE = "NONE",
  FORBIDDEN = "FORBIDDEN",
  SPRINT_NOT_FOUND = "SPRINT_NOT_FOUND",
  TASK_NOT_FOUND = "TASK_NOT_FOUND",
  FILE_NOT_FOUND = "FILE_NOT_FOUND",
  INVALID_INPUT = "INVALID_INPUT",
  INTERNAL_ERROR = "INTERNAL_ERROR",
  USER_NOT_FOUND = "USER_NOT_FOUND",
  PROJECT_NOT_FOUND = "PROJECT_NOT_FOUND"
}